mvp for hooking up the database complete

+++ app.js
@@ -3,7 +3,8 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
-const date = require(__dirname + "/date.js");
+const _ = require("lodash");
+
 
 const app = express();
 
@@ -63,10 +64,7 @@ app.get("/", function (req, res) {
     else {
       res.render("list", { listTitle: "Today", newListItems: foundItems });
     }
-
   })
-
-
 });
 
 app.post("/", function (req, res) {
@@ -77,7 +75,6 @@ app.post("/", function (req, res) {
   const item = new Item({
     name: itemName
   })
-
 
   if (listName === "Today") {
     item.save()
@@ -98,19 +95,28 @@ app.post("/", function (req, res) {
 
 app.post("/delete", (req, res) => {
   const checkedItemId = req.body.checkbox;
-  Item.findByIdAndRemove(checkedItemId, (err) => {
-    if (err) {
-      console.log(err)
-    }
-    else {
-      console.log("Item was successfully deleted")
-    }
-  })
-  res.redirect("/")
+  const listName = req.body.listName;
+
+  if (listName === "Today") {
+    Item.findByIdAndRemove(checkedItemId, (err) => {
+      if (err) {
+        console.log(err)
+      }
+      else {
+        List.findOneAndUpdate({ name: listName }, { $pull: { items: { _id: checkedItemId } } }, function (err, foundList) {
+          if (!err) {
+            res.redirect("/" + listName)
+          }
+        })
+      }
+    })
+    res.redirect("/")
+  }
+
 })
 
 app.get("/:customListName", function (req, res) {
-  const customListName = req.params.customListName;
+  const customListName = _.capitalize(req.params.customListName);
 
   List.findOne({ name: customListName }, function (err, foundList) {
     if (!err) {
